# shared_memory/api-contract.yaml
openapi: 3.1.0
info:
  title: Mực API
  version: 1.0.0
  description: Public catalog, personal reading state, source ingestion and BYOK recommendation API.
servers:
  - url: /
paths:
  /api/catalog:
    get:
      operationId: listCatalog
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 120 }, description: Title, alias, or author query, example: "nữ chính trả thù" }
        - { name: include, in: query, schema: { type: array, items: { type: string } }, description: Genre or tag slugs to include, example: [fantasy, revenge] }
        - { name: exclude, in: query, schema: { type: array, items: { type: string } }, description: Genre or tag slugs to exclude, example: [harem] }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/StoryStatus' }, description: Publication status, example: ongoing }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, latest, rating, popular, shortest], default: relevance }, description: Result ordering, example: latest }
        - { name: cursor, in: query, schema: { type: string }, description: Opaque pagination cursor, example: "c_01J..." }
      responses:
        '200':
          description: Catalog page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogResponse' }
        default: { $ref: '#/components/responses/StandardError' }
  /api/stories/{slug}:
    get:
      operationId: getStory
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string, pattern: '^[a-z0-9-]+$' }, description: Canonical story slug, example: "hoa-son-tai-xuat" }
      responses:
        '200':
          description: Story details, sources, score and chapters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StoryDetail' }
        '404': { $ref: '#/components/responses/StandardError' }
  /api/progress:
    put:
      operationId: upsertProgress
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProgressWrite' }
      responses:
        '200':
          description: Saved progress
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Progress' }
        default: { $ref: '#/components/responses/StandardError' }
  /api/library:
    get:
      operationId: listLibrary
      responses:
        '200':
          description: Current user's library
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/LibraryEntry' } }
                additionalProperties: false
        default: { $ref: '#/components/responses/StandardError' }
    put:
      operationId: upsertLibraryEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storyId, status, followed]
              properties:
                storyId: { type: string, description: Story ID, example: "story_01J" }
                status: { type: string, enum: [reading, planned, completed, paused, dropped], description: Shelf status, example: reading }
                followed: { type: boolean, description: Notify on new chapters, example: true }
              additionalProperties: false
      responses:
        '200':
          description: Updated entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryEntry' }
        default: { $ref: '#/components/responses/StandardError' }
  /api/download-manifest/{chapterId}:
    get:
      operationId: getDownloadManifest
      parameters:
        - { name: chapterId, in: path, required: true, schema: { type: string }, description: Chapter ID, example: "chapter_47" }
      responses:
        '200':
          description: Validated chapter download manifest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DownloadManifest' }
        default: { $ref: '#/components/responses/StandardError' }
  /api/ai/recommend:
    post:
      operationId: recommendWithUserKey
      description: Proxies a constrained recommendation request without storing the supplied API key.
      parameters:
        - { name: X-AI-Provider, in: header, required: true, schema: { type: string, enum: [openai-compatible, anthropic, gemini] }, description: Provider adapter, example: openai-compatible }
        - { name: X-AI-Key, in: header, required: true, schema: { type: string, minLength: 8 }, description: Session-only user key; never persisted, example: "sk-..." }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, candidateIds, model]
              properties:
                query: { type: string, minLength: 1, maxLength: 500, description: User's recommendation request, example: "Tìm truyện trả thù ít romance" }
                candidateIds: { type: array, minItems: 1, maxItems: 30, items: { type: string }, description: Server-known candidates only, example: ["story_01J"] }
                model: { type: string, minLength: 1, maxLength: 80, description: Developer-approved model name, example: "gpt-4.1-mini" }
              additionalProperties: false
      responses:
        '200':
          description: Ranked candidates with explanations
          content:
            application/json:
              schema:
                type: object
                required: [recommendations]
                properties:
                  recommendations:
                    type: array
                    items:
                      type: object
                      required: [storyId, reason]
                      properties:
                        storyId: { type: string, description: Existing candidate story ID, example: "story_01J" }
                        reason: { type: string, description: Short grounded explanation, example: "Trùng tag trả thù và nhịp nhanh." }
                      additionalProperties: false
                additionalProperties: false
        default: { $ref: '#/components/responses/StandardError' }
  /api/admin/ingest:
    post:
      operationId: runIncrementalIngest
      parameters:
        - { name: X-Ingest-Token, in: header, required: true, schema: { type: string, minLength: 16 }, description: Hosting secret, example: "runtime-secret" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, mode]
              properties:
                source: { type: string, enum: [otruyen, opds, anilist], description: Registered connector slug, example: otruyen }
                mode: { type: string, enum: [incremental, refresh], description: Sync mode, example: incremental }
                cursor: { type: [string, 'null'], description: Optional resume cursor, example: null }
              additionalProperties: false
      responses:
        '202':
          description: Sync accepted
          content:
            application/json:
              schema:
                type: object
                required: [runId, status]
                properties:
                  runId: { type: string, description: Sync run ID, example: "sync_01J" }
                  status: { type: string, const: accepted, description: Accepted state, example: accepted }
                additionalProperties: false
        default: { $ref: '#/components/responses/StandardError' }
components:
  schemas:
    StoryStatus:
      type: string
      enum: [ongoing, completed, hiatus, cancelled]
      description: Normalized publication status
      example: ongoing
    StorySummary:
      type: object
      required: [id, slug, title, coverUrl, status, genres, latestChapter, score]
      properties:
        id: { type: string, description: Stable story ID, example: "story_01J" }
        slug: { type: string, description: URL slug, example: "hoa-son-tai-xuat" }
        title: { type: string, description: Display title, example: "Hoa Sơn Tái Xuất" }
        coverUrl: { type: [string, 'null'], format: uri, description: Source-attributed cover URL or null, example: "https://example.org/cover.jpg" }
        status: { $ref: '#/components/schemas/StoryStatus' }
        genres: { type: array, items: { type: string }, description: Normalized genre labels, example: [Action, Fantasy] }
        latestChapter: { type: [number, 'null'], description: Latest numeric chapter when known, example: 147 }
        score: { $ref: '#/components/schemas/AggregateScore' }
      additionalProperties: false
    CatalogResponse:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StorySummary' } }
        nextCursor: { type: [string, 'null'], description: Opaque next cursor, example: null }
      additionalProperties: false
    AggregateScore:
      type: object
      required: [value, confidence, sourceCount, voteCount, computedAt]
      properties:
        value: { type: [number, 'null'], minimum: 0, maximum: 5, description: Bayesian aggregate or null, example: 4.6 }
        confidence: { type: string, enum: [insufficient, low, medium, high], description: Evidence confidence, example: high }
        sourceCount: { type: integer, minimum: 0, description: Number of score sources, example: 3 }
        voteCount: { type: integer, minimum: 0, description: Total source votes, example: 12430 }
        computedAt: { type: string, format: date-time, description: Score computation time, example: "2026-07-23T09:00:00Z" }
      additionalProperties: false
    StoryDetail:
      allOf:
        - { $ref: '#/components/schemas/StorySummary' }
        - type: object
          required: [synopsis, author, chapters, sources]
          properties:
            synopsis: { type: string, description: Story synopsis, example: "Một kiếm sĩ trở lại..." }
            author: { type: [string, 'null'], description: Display author, example: "Biga" }
            chapters: { type: array, items: { $ref: '#/components/schemas/Chapter' } }
            sources: { type: array, items: { $ref: '#/components/schemas/SourceEvidence' } }
          additionalProperties: false
    Chapter:
      type: object
      required: [id, number, title, language, pageCount, publishedAt, sourceUrl]
      properties:
        id: { type: string, description: Chapter ID, example: "chapter_47" }
        number: { type: number, description: Chapter number, example: 47 }
        title: { type: string, description: Optional source title, example: "Trở về" }
        language: { type: string, description: BCP-47 language code, example: vi }
        pageCount: { type: integer, minimum: 0, description: Page count, example: 28 }
        publishedAt: { type: string, format: date-time, description: Source publication time, example: "2026-07-22T12:00:00Z" }
        sourceUrl: { type: string, format: uri, description: Original chapter link, example: "https://example.org/chapter/47" }
      additionalProperties: false
    SourceEvidence:
      type: object
      required: [name, url, lastSyncAt, licenseMode]
      properties:
        name: { type: string, description: Source display name, example: OTruyen }
        url: { type: string, format: uri, description: Original item URL, example: "https://otruyen.cc/truyen-tranh/x" }
        lastSyncAt: { type: string, format: date-time, description: Last successful sync, example: "2026-07-23T08:00:00Z" }
        licenseMode: { type: string, enum: [api, opds, feed, link-only], description: Allowed connector mode, example: api }
      additionalProperties: false
    ProgressWrite:
      type: object
      required: [storyId, chapterId, page, progress, idempotencyKey]
      properties:
        storyId: { type: string, description: Story ID, example: "story_01J" }
        chapterId: { type: string, description: Chapter ID, example: "chapter_47" }
        page: { type: integer, minimum: 0, description: Zero-based page, example: 12 }
        progress: { type: number, minimum: 0, maximum: 1, description: Normalized progress, example: 0.43 }
        idempotencyKey: { type: string, minLength: 8, maxLength: 128, description: Unique write key, example: "prog_story_01J_47_12" }
      additionalProperties: false
    Progress:
      allOf:
        - { $ref: '#/components/schemas/ProgressWrite' }
        - type: object
          required: [updatedAt]
          properties:
            updatedAt: { type: string, format: date-time, description: Server update time, example: "2026-07-23T09:00:00Z" }
          additionalProperties: false
    LibraryEntry:
      type: object
      required: [story, status, followed, updatedAt]
      properties:
        story: { $ref: '#/components/schemas/StorySummary' }
        status: { type: string, enum: [reading, planned, completed, paused, dropped], description: Shelf state, example: reading }
        followed: { type: boolean, description: New chapter follow flag, example: true }
        updatedAt: { type: string, format: date-time, description: Last update, example: "2026-07-23T09:00:00Z" }
      additionalProperties: false
    DownloadManifest:
      type: object
      required: [chapterId, version, estimatedBytes, pages, sourceUrl]
      properties:
        chapterId: { type: string, description: Chapter ID, example: "chapter_47" }
        version: { type: string, description: Cache invalidation version, example: "v1-20260723" }
        estimatedBytes: { type: integer, minimum: 0, description: Estimated download bytes, example: 24500000 }
        pages: { type: array, items: { type: string, format: uri }, description: Validated image URLs, example: ["https://example.org/p1.webp"] }
        sourceUrl: { type: string, format: uri, description: Original chapter URL, example: "https://example.org/chapter/47" }
      additionalProperties: false
  responses:
    StandardError:
      description: Standard error response
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, description: Human-safe error, example: "Invalid request" }
              code: { type: string, description: Stable machine code, example: INVALID_REQUEST }
              details: { type: [object, 'null'], description: Non-sensitive structured details, additionalProperties: true, example: null }
            additionalProperties: false